# backend/docs/swagger.yaml
basePath: /
definitions:
...
  main.TransactionSplit:
    properties:
      amount:
        type: number
      category_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      notes:
        type: string
      transaction_id:
        type: string
      updated_at:
        type: string
    type: object
  main.splitRequest:
    properties:
      splits:
        items:
          properties:
            amount:
              type: number
            category_id:
              type: string
            notes:
              type: string
          type: object
        type: array
    type: object
host: localhost:8081
info:
  contact: {}
  description: API for the Joint Analysis expense tracking application
  title: Joint Analysis API
  version: "1.0"
paths:
...
  /api/transactions/{id}/splits:
    get:
      description: Retrieve split rows for a transaction. Returns a default single
        split derived from transaction category/amount when no explicit splits are
        stored.
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of splits
          schema:
            items:
              $ref: '#/definitions/main.TransactionSplit'
            type: array
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Transaction not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get transaction splits
      tags:
      - transactions
    put:
      consumes:
      - application/json
      description: Replace all split rows for a transaction and sync transaction category
        to the first split for backward compatibility.
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      - description: Split rows
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.splitRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated splits
          schema:
            items:
              $ref: '#/definitions/main.TransactionSplit'
            type: array
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Transaction not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Replace transaction splits
      tags:
      - transactions
...
swagger: "2.0"